beckend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
openai
sentence-transformers
faiss-cpu
python-dotenv
google.generativeai
langchain-core
langchain-huggingface
langchin-community
langchain
